python-backend/requirements.txt
sanic>=23.12.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
ultralytics>=8.0.0
onnxruntime-gpu>=1.16.0;platform_system!="Darwin"
onnxruntime>=1.16.0;platform_system=="Darwin"
psutil>=5.9.0
rembg>=2.0.0
timm>=1.0.0
mobile-sam @ git+https://github.com/ChaoningZhang/MobileSAM.git
